+++ src/builtin/variable.rs
@@ -1,57 +1,49 @@
-use super::super::{
-  ArgLen::Exactly,
+use crate::{
+  Arity::Exactly,
   Bind::{self, Lit, Var},
-  ErrOR, FuncInfo, Json, Jsonpiler, ScopeInfo, err, mn,
-  utility::{get_bool_str, get_int_str},
-  validate_type,
+  ErrOR, FuncInfo, Json, Jsonpiler, ScopeInfo, built_in, err, mn, take_arg,
 };
 impl Jsonpiler {
-  pub(crate) fn variable(&mut self) {
-    let common = (false, false);
-    let sp_scope = (true, true);
-    self.register("global", common, Jsonpiler::assign_global, Exactly(2));
-    self.register("=", common, Jsonpiler::assign_local, Exactly(2));
-    self.register("$", common, Jsonpiler::reference, Exactly(1));
-    self.register("scope", sp_scope, Jsonpiler::scope, Exactly(1));
-  }
-}
-#[expect(clippy::single_call_fn, reason = "")]
-impl Jsonpiler {
-  fn assign(&mut self,  func: &mut FuncInfo, scope: &mut ScopeInfo, is_global: bool) -> ErrOR<Json> {
-    let json1 = func.arg()?;
-    let variable = validate_type!(self, func, 1, json1, Json::String(Lit(x)) => x, "String");
+  fn assign(&mut self, func: &mut FuncInfo, scope: &mut ScopeInfo, is_global: bool) -> ErrOR<Json> {
+    let (variable, pos) = take_arg!(self, func, 1, "String", Json::String(Lit(x)) => x);
     let json2 = func.arg()?;
     let value = match json2.value {
       Json::Function(asm_func) => {
         if self.builtin.contains_key(&variable) {
-          return err!(self, json1.pos, "Name conflict with a built-in function.");
+          return err!(self, pos, "Name conflict with a built-in function.");
         }
         Json::Function(asm_func)
       }
       Json::String(Lit(st)) => Json::String(Var(self.global_str(&st)?)),
-      Json::String(Var(_)) if is_global => {
-        return err!(self, json2.pos, "Local string cannot be assigned to a global variable.");
+      Json::String(Var(str_label)) => {
+        let label = if is_global { self.get_bss(8) } else { scope.local(8) }?;
+        scope.body.push(mn!("lea", label, str_label));
+        Json::Int(Var(label))
       }
-      var @ Json::String(Var(_)) => var,
       Json::Null => Json::Null,
       Json::Int(Lit(int)) if is_global => Json::Int(Var(self.global_num(int)?)),
       Json::Int(int) => {
-        let label = if is_global { self.get_bss(8) } else { scope.get_local(8) }?;
-        let int_str = get_int_str(&int, func);
+        let label = if is_global { self.get_bss(8) } else { scope.local(8) }?;
+        let int_str = match int {
+          Lit(l_int) => l_int.to_string(),
+          Var(int_label) => format!("{int_label}"),
+        };
         scope.body.push(mn!("mov", label, int_str));
         Json::Int(Var(label))
       }
+      Json::Bool(Lit(boolean)) if is_global => Json::Bool(Var(self.global_bool(boolean)?)),
       Json::Bool(boolean) => {
-        let label = if is_global { self.get_bss(1) } else { scope.get_local(1) }?;
-        let bool_str = get_bool_str(&boolean, func);
+        let label = if is_global { self.get_bss(1) } else { scope.local(1) }?;
+        let bool_str = match boolean {
+          Lit(l_bool) => if l_bool { "0xFF" } else { "0x00" }.to_owned(),
+          Var(bool_label) => format!("{bool_label}"),
+        };
         scope.body.push(mn!("mov", label, bool_str));
         Json::Bool(Var(label))
       }
-      Json::Float(Lit(float)) if is_global => {
-        Json::Float(Var(self.global_num(float.to_bits())?))
-      }
+      Json::Float(Lit(float)) if is_global => Json::Float(Var(self.global_num(float.to_bits())?)),
       Json::Float(float) => {
-        let label = if is_global { self.get_bss(8) } else { scope.get_local(8) }?;
+        let label = if is_global { self.get_bss(8) } else { scope.local(8) }?;
         let float_str = match float {
           Bind::Lit(l_float) => format!("{:#016x}", l_float.to_bits()),
           Bind::Var(float_label) => float_label.sched_free_2str(func),
@@ -60,11 +52,16 @@ impl Jsonpiler {
         Json::Float(Var(label))
       }
       Json::Array(_) | Json::Object(_) => {
-        return self.parser.typ_err(
-          2,
-          &func.name,
-          "that supports assignment (excluding Array and Object)",
-          &json2,
+        return Err(
+          self
+            .parser
+            .type_err(
+              2,
+              &func.name,
+              "that supports assignment (excluding Array and Object)",
+              &json2,
+            )
+            .into(),
         );
       }
     };
@@ -76,26 +73,27 @@ impl Jsonpiler {
     .insert(variable, value)
     .is_some()
     {
-      return err!(self, json1.pos, "Reassignment may not be possible in some scope.");
+      return err!(self, pos, "Reassignment may not be possible in some scope.");
     }
     Ok(Json::Null)
   }
-  fn assign_global(&mut self, func: &mut FuncInfo, scope: &mut ScopeInfo) -> ErrOR<Json> {
+}
+built_in! {self, func, scope, variable;
+  assign_global => {"global", COMMON, Exactly(2), {
     self.assign(func, scope, true)
-  }
-  fn assign_local(&mut self, func: &mut FuncInfo, scope: &mut ScopeInfo) -> ErrOR<Json> {
+  }},
+  assign_local =>{ "=", COMMON, Exactly(2), {
     self.assign(func, scope, false)
-  }
-  fn reference(&mut self, func: &mut FuncInfo, scope: &mut ScopeInfo) -> ErrOR<Json> {
-    let json = func.arg()?;
-    let var_name =
-      validate_type!(self, func, 1, json, Json::String(Lit(x)) => x, "String(Literal)");
-    match self.get_var(&var_name, scope){Some(var) => Ok(var), None => err!(self, json.pos, "Undefined variables: `{var_name}`")}
-  }
-  fn scope(&mut self, func: &mut FuncInfo, scope: &mut ScopeInfo) -> ErrOR<Json> {
-    let json = func.arg()?;
-    let mut object =
-      validate_type!(self, func, 1, json, Json::Object(Lit(x)) => x, "Object(Literal)");
-    self.eval_object(&mut object, scope)
-  }
+  }},
+  reference => {"$", COMMON, Exactly(1), {
+    let (var_name, pos) = take_arg!(self, func, 1, "String(Literal)", Json::String(Lit(x)) => x);
+    match self.get_var(&var_name, scope) {
+      Some(var) => Ok(var),
+      None => err!(self, pos, "Undefined variables: `{var_name}`"),
+    }
+  }},
+  scope => {"scope", SP_SCOPE, Exactly(1), {
+    let (object, object_pos) = take_arg!(self, func, 1, "Sequence", Json::Object(Lit(x)) => x);
+    self.eval_object(object, object_pos, scope)
+  }}
 }