+++ src/func_info.rs
@@ -1,7 +1,7 @@
 use crate::{ErrOR, FuncInfo, Json, Label, VarKind::Tmp, WithPos};
 impl FuncInfo {
   pub fn arg(&mut self) -> ErrOR<WithPos<Json>> {
-    self.args.pop_front().ok_or("InternalError: Invalid argument reference".into())
+    self.args.next().ok_or("InternalError: Invalid argument reference".into())
   }
   pub fn sched_free_tmp(&mut self, label: &Label) {
     if label.kind == Tmp {