+++ src/macros.rs
@@ -1,48 +1,24 @@
 #[macro_export]
 #[doc(hidden)]
-macro_rules! add {
-  ($op1: expr, $op2: expr) => {
-    $op1.checked_add($op2).ok_or("InternalError: Overflow occurred")
-  };
-}
-#[macro_export]
-#[doc(hidden)]
-macro_rules! def_mod_and_register {
-  ($($name:ident),* $(,)?) => {
-    $(
-      mod $name;
-    )*
-    use super::{
-  ArgLen,
-  Bind::{Lit, Var},
-  Builtin, ErrOR, FuncInfo, JFunc, Json, Jsonpiler, ScopeInfo, WithPos, err, mn,
-  mn_write,
-  utility::{get_int_str_without_free, imp_call},
-};
-use core::mem::take;
-use std::{
-  collections::VecDeque,
-  io::Write as _,
-};
-    impl Jsonpiler {
-      pub(crate) fn register_all(&mut self) {
-        $(
-          self.$name();
-        )*
-      }
+macro_rules! return_if {
+  ($self:ident, $ch:expr, $pos:ident, $value:expr) => {
+    if $self.peek() == $ch {
+      $self.pos.offset += 1;
+      $pos.extend_to($self.pos.offset);
+      return Ok(WithPos { $pos, value: $value });
     }
   };
 }
 #[macro_export]
 #[doc(hidden)]
 macro_rules! parse_err {
-  ($self:ident, $pos:expr, $($arg:tt)*) => {Err($self.fmt_err(&format!($($arg)*), &$pos).into())};
-  ($self:ident, $($arg:tt)*) => {Err($self.fmt_err(&format!($($arg)*), &$self.pos).into())};
+  ($self:ident, $pos:expr, $($arg:tt)*) => {Err($self.fmt_err(&format!($($arg)*), $pos).into())};
+  ($self:ident, $($arg:tt)*) => {Err($self.fmt_err(&format!($($arg)*), $self.pos).into())};
 }
 #[macro_export]
 #[doc(hidden)]
 macro_rules! err {
-  ($self:ident, $pos:expr, $($arg:tt)*) => {Err($self.parser.fmt_err(&format!($($arg)*), &$pos).into())};
+  ($self:ident, $pos:expr, $($arg:tt)*) => {Err($self.parser.fmt_err(&format!($($arg)*), $pos).into())};
 }
 #[macro_export]
 #[doc(hidden)]
@@ -78,31 +54,65 @@ macro_rules! mn_write {
 }
 #[macro_export]
 #[doc(hidden)]
+macro_rules! add {
+  ($op1:expr, $op2:expr) => {
+    $op1.checked_add($op2).ok_or("InternalError: Overflow occurred")
+  };
+}
+#[macro_export]
+#[doc(hidden)]
 macro_rules! sub {
-  ($op1: expr, $op2: expr) => {
+  ($op1:expr, $op2:expr) => {
     $op1.checked_sub($op2).ok_or("InternalError: Underflow occurred")
   };
 }
 #[macro_export]
 #[doc(hidden)]
 macro_rules! warn {
-  ($self:ident, $pos:expr, $($arg:tt)*) => {println!("Warning: {}", $self.parser.fmt_err(&format!($($arg)*), &$pos))};
+  ($self:ident, $pos:expr, $($arg:tt)*) => {println!("Warning: {}", $self.parser.fmt_err(&format!($($arg)*), $pos))};
 }
 #[macro_export]
 #[doc(hidden)]
-macro_rules! validate_type {
+macro_rules! take_arg {
   (
     $self:ident,
     $func:expr,
-    $ord:expr,
-    $val:expr,
-    $pat:pat => $body:expr,
-    $expected:literal
-  ) => {
-    if let $pat = $val.value {
-      $body
+    $nth:expr,
+    $expected:literal,
+    $pat:pat => $body:expr
+  ) => {{
+    let arg = $func.arg()?;
+    if let $pat = arg.value {
+      ($body, arg.pos)
     } else {
-      return $self.parser.typ_err($ord, &$func.name, $expected, &$val);
+      return Err($self.parser.type_err($nth, &$func.name, $expected, &arg).into());
     }
-  };
+  }};
+}
+#[macro_export]
+#[doc(hidden)]
+macro_rules! built_in {
+  (
+    $self:ident, $func:ident, $scope:ident, $register_fn:ident;
+    $( $name:ident => { $key:literal, $attrs:tt, $arity:expr, $block:block } ),+ $(,)?
+  ) => {
+    impl Jsonpiler {
+      pub(crate) fn $register_fn(&mut self) {
+      $(
+        self.register($key, Jsonpiler::$attrs, Jsonpiler::$name, $arity);
+      )+
+      }
+    }
+    #[allow(clippy::allow_attributes, clippy::single_call_fn, clippy::unnecessary_wraps)]
+    impl Jsonpiler {
+    $(
+      fn $name(
+        &mut $self,
+        $func: &mut FuncInfo,
+        $scope: &mut ScopeInfo
+      ) -> ErrOR<Json> {
+        $block
+      }
+    )+
+  }};
 }